requirements.txt
fastapi
uvicorn
python-dotenv
paddleocr
openai
supabase
paddlepaddle
